+++ tests/script_engine/test_script.py
@@ -10,7 +10,6 @@
 """Tests for the `btclib.script.engine` module."""
 
 import json
-import warnings
 from os import path
 
 import pytest
@@ -87,15 +86,11 @@ def test_script() -> None:
         flags = x[i + 2]
         result = x[i + 3] == "OK"
 
-        with warnings.catch_warnings():
-            warnings.simplefilter("ignore")
-            if result:
+        if result:
+            test(stack, amount, script_sig_str, script_pub_key_str, flags, result)
+        else:
+            with pytest.raises(Exception):
                 test(stack, amount, script_sig_str, script_pub_key_str, flags, result)
-            else:
-                with pytest.raises(Exception):
-                    test(
-                        stack, amount, script_sig_str, script_pub_key_str, flags, result
-                    )
 
 
 def test_script_error_says_what_and_where() -> None: